news_crawler/requirements.txt
firebase-admin==6.6.0
feedparser==6.0.11
google-generativeai==0.8.3
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.31.0
beautifulsoup4==4.12.3
